backend/requirements.txt

# FastAPI and server
fastapi
uvicorn[standard]
python-multipart
pydantic
#pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
torch
transformers==4.35.0
scikit-learn
numpy

# NLP Libraries
nltk
textblob

# Data processing
joblib==1.3.2

# Environment and utilities
python-dotenv==1.0.0